feat: Service Impl

+++ back/src/main/java/com/teamrocket/projetdevop/ivvqproject/repositories/ShoppingCartRepository.java
@@ -3,5 +3,5 @@ package com.teamrocket.projetdevop.ivvqproject.repositories;
 import com.teamrocket.projetdevop.ivvqproject.domain.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
 }
